feat: home as drafts hub, camera close UX, light/dark mode, upload destination
- Move all drafts to home: Welcome to Pulse shows draft list, Import, Export
- Remove Shorts/Drafts nav from camera; + opens camera with new draft every time
- No auto-load last draft when opening camera without draftId (useDraftManager)
- Close button: single action (no Start New/Start Over/Close menu); upload mode
  closes to (tabs) to avoid white screen after QR deeplink
- Home: light/dark mode (Colors surface/border/secondaryText/selection), delete
  icon (MaterialIcons delete-outline), upload drafts show full server URL below
  segment/timestamp; Export only when drafts exist
- Remove (camera)/drafts.tsx; simplify CloseButton and UploadCloseButton to
  onClose only
